# core/services/versionchooser/openapi/versionchooser.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: BlueOS Version Chooser
  license:
    name: MIT
servers:
  - url: http://localhost:8081/v1.0

paths:
  /version/current:
    get:
      operationId: main.get_version
      summary: Return the current running version of BlueOS
      responses:
        '200':
          description: The current running version of BlueOS
    post:
      operationId: main.set_version
      summary: Sets the current version of BlueOS to a new tag
      responses:
        '200':
          description: Successfully set a new version
        '400':
          description: Attempted to set an unavailable version
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                repository:
                  type: string
                  example: bluerobotics/blueos-core
                tag:
                  type: string
                  example: master

  /bootstrap/current:
    get:
      operationId: main.get_bootstrap_version
      summary: Return the current running version of BlueOS-bootstrap
      responses:
        '200':
          description: The current running version of BlueOS-bootstrap
    post:
      operationId: main.set_bootstrap_version
      summary: Sets the current version of BlueOS-bootstrap to a new tag
      responses:
        '200':
          description: Successfully set a new version
        '400':
          description: Attempted to set an unavailable version
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                tag:
                  type: string
                  example: master

  /version/restart:
    post:
      operationId: main.restart
      responses:
        '200':
          description: Restart the currently running docker container

  /version/delete:
    delete:
      operationId: main.delete_version
      summary: Delete the selected version of BlueOS
      responses:
        '200':
          description: Version as deleted successfully
        '404':
          description: Version does not exist
        '403':
          description: Version can not be deleted
        '500':
          description: Internal error while trying to delete the file
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                repository:
                  type: string
                  example: bluerobotics/blueos-core
                tag:
                  type: string
                  example: master

  /version/available/local:
    get:
      operationId: main.get_available_local_versions
      summary: Returns available local versions
      responses:
        '200':
          description: A list of versions available locally
          content:
            application/json:
              schema:
                type: object
                properties:
                  local:
                    type: array
                    items:
                      $ref: '#/components/schemas/Version'
                  error:
                    type: string
                    example: "Unable to fetch sha."

  /version/available/{repository}/{image}:
    get:
      operationId: main.get_available_versions
      summary: Returns available versions, both locally and in dockerhub
      parameters:
        - in: path
          name: repository
          schema:
            type: string
          x-nullable: true
          required: true
        - name: image
          in: path
          schema:
            type: string
          required: true
          example: blueos-core
      responses:
        '200':
          description: A list of versions available locally and on the remote(s)
          content:
            application/json:
              schema:
                type: object
                properties:
                  remote:
                    type: array
                    items:
                      $ref: '#/components/schemas/Version'
                  local:
                    type: array
                    items:
                      $ref: '#/components/schemas/Version'
                  error:
                    type: string
                    example: "Unable to fetch sha."

  /version/pull/:
    post:
      operationId: main.pull_version
      summary: Pulls a version from dockerhub

      responses:
        '200':
          description: Successfully set a new version
        '400':
          description: Attempted to set an unavailable version
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                repository:
                  type: string
                  example: bluerobotics/blueos-core
                tag:
                  type: string
                  example: master

  /version/load/:
    post:
      operationId: main.load
      summary: Load a docker tar file
      responses:
        '200':
          description: Successfully loaded the file
        '400':
          description: Error while processing file
        '500':
          description: Internal server error while processing docker import
        '501':
          description: Failed to handle docker result
      requestBody:
        content:
          application/x-tar:
            schema:
              type: string
              format: binary

components:
  schemas:
    Version:
      type: object
      properties:
        image:
          type: string
          example: bluerobotics/blueos-core
        tag:
          type: string
          example: master